people is now a functional component
